==> scripts/00_run_all_code.R <==
library(tictoc)

tic()
source("01_simulate_data.R")
source("02_naive.R")
source("03_complete_case.R")
source("04_01_snipe_sandwich.R")
source("04_02_snipe_bootstrap.R")
source("05_01_split_datasets.R")
source("05_02_create_imputations.R")
source("05_02_01_create_full_stochastic_imputations.R")
source("05_03_pool_results.R")
source("07_true_model.R")
source("08_compile_results.R")
toc()
